requirements.txt
pyyaml==6.0.1
numpy
torch==1.11.0
torchtext==0.12.0
torchdata==0.3.0
